lambdaアップロード用に修正
