+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "traindelay",
   "version": "1.0.0",
-  "main": "app.js",
+  "main": "index.js",
   "license": "MIT",
   "dependencies": {
     "axios": "^0.21.1",
